Move create_sparse_query_vector function to sparse_ann_query_basic.rs
